python/llama-index-server/pyproject.toml
[project]
name = "llama-index-server"
version = "0.1.20"
description = "llama-index fastapi server"
readme = "README.md"
license = "MIT"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi[standard]>=0.115.11,<1.0.0",
    "cachetools>=5.5.2,<6.0.0",
    "requests>=2.32.3,<3.0.0",
    "pydantic-settings>=2.8.1,<3.0.0",
    "llama-index-core>=0.12.28,<1.0.0",
    "llama-index-readers-file>=0.4.6,<1.0.0",
    "llama-index-indices-managed-llama-cloud>=0.6.3,<1.0.0",
]


[tool.codespell]
check-filenames = true
check-hidden = true
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true
python_version = "3.10"


[build-system]
requires = [ "hatchling>=1.24" ]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "llama-index-llms-openai>=0.3.38",
    "black[jupyter]<=23.9.1,>=23.7.0",
    "codespell[toml]>=2.2.6",
    "e2b-code-interpreter>=1.1.1,<2.0.0",
    "ipython==8.10.0",
    "jupyter>=1.0.0,<2.0.0",
    "markdown>=3.7,<4.0",
    "mypy==1.15.0",
    "pre-commit==3.2.0",
    "pylint==2.15.10",
    "pytest>=8.3.5,<9.0.0",
    "pytest-asyncio>=0.25.3,<1.0.0",
    "pytest-mock==3.11.1",
    "ruff==0.0.292",
    "tree-sitter-languages>=1.8.0,<2.0.0",
    "types-Deprecated>=0.1.0",
    "types-PyYAML>=6.0.12.12,<7.0.0.0",
    "types-protobuf>=4.24.0.4,<5.0.0.0",
    "types-redis==4.5.5.0",
    "types-requests==2.28.11.8",
    "types-setuptools==67.1.0.0",
    "xhtml2pdf>=0.2.17,<1.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "llama-cloud>=0.1.17,<1.0.0",
]

[tool.hatch.build]
packages = ["llama_index/"]
artifacts = ["llama_index/server/resources"]
